pyproject.toml
[project]
name = "synapse-core"
version = "0.1.0"
description = "Turn your files into answers — local RAG ingestion pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb",
    "sentence-transformers",
    "pypdf",
    "python-docx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
